Things to add in/edit
* computational/baseline sequential algorithm
The time complexity of modelling the reversal as in models,  is mainly dependent on the resolution of the model. For high fidelity simulation we need to increase the problem size N , to simulate 300000 ms .
We reduce the complexity by parallelising in two models:

##SIMT parallelization 

Many-core througput oriented GPU architectures:  Here we program first in SIMT for a GPU on Odyssey that gives us the first level of throughput improvement as shown in figure ...

##SIMT parallelization P100 many-core
We further increase the througput by programming in SIMT but for a higher throughput P100 GPU architectures. This enables us to perform more simulations....

##Hybrid SPMD parallelization
openacc + MPI (bcast) 

* put thoroughplot with parallelization 
* time complexity 
* bcast for mpi